Renaming and Refactoring MLP envmodel for robotinhtm
