+++ src/mlpro/sl/pool/afct/mlpafctrans_robotinhtm.py
@@ -1,7 +1,7 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro.sl.afct
-## -- Module  : mlpafctrans
+## -- Module  : mlpafctrans_robotinhtm
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
